+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Auto-hotstaging support in StageController
 ### Changed
 - **AscentPEG** module overhaul (renamed to **AscentPVG**)
 ### Fixed